Use size to calculate buffer size instead of array size in sorted_check. Hand over numberOfValues to external_sort instead of file size. Sort ascending in queue. Flush temp file after write.

+++ src/assignment1/sorted_check.h
@@ -7,6 +7,6 @@
 
 #include <stddef.h>
 
-bool check_sorting(int fd, size_t size);
+bool check_sorting(int fd, uint64_t size);
 
 #endif //PROJECT_SORTED_CHECK_H
